requirements.txt
openai
chromadb
pypdf2
streamlit
python-dotenv
sentence-transformers
reportlab
fastapi
uvicorn
